update :: student response
