+++ bitgouel-api/src/main/kotlin/team/msg/domain/club/presentation/data/response/ClubResponse.kt
@@ -1,9 +1,6 @@
 package team.msg.domain.club.presentation.data.response
 
 import team.msg.domain.club.model.Club
-import team.msg.domain.student.model.Student
-import team.msg.domain.user.enums.Authority
-import java.util.*
 
 data class ClubResponse(
     val id: Long,
@@ -22,14 +19,6 @@ data class ClubResponse(
             highSchoolName = club.school.highSchool.schoolName,
             headCount = headCount
         )
-
-        fun listOfStudent(students: List<Student>) = students.map {
-            StudentResponse(
-                id = it.user!!.id,
-                name = it.user!!.name,
-                authority = it.user!!.authority
-            )
-        }
     }
 }
 
@@ -41,14 +30,4 @@ data class ClubDetailsResponse(
     val clubName: String,
     val highSchoolName: String,
     val headCount: Int
-)
-
-data class StudentResponse(
-    val id: UUID,
-    val name: String,
-    val authority: Authority
-)
-
-data class AllStudentsResponse(
-    val students: List<StudentResponse>
 )
