+++ bitgouel-api/src/main/kotlin/team/msg/domain/club/service/ClubService.kt
@@ -2,8 +2,8 @@ package team.msg.domain.club.service
 
 import team.msg.domain.club.presentation.data.response.AllClubResponse
 import team.msg.domain.club.presentation.data.response.ClubDetailsResponse
-import team.msg.domain.club.presentation.data.response.AllStudentsResponse
 import team.msg.domain.school.enums.HighSchool
+import team.msg.domain.student.presentation.data.response.AllStudentsResponse
 
 interface ClubService {
     fun queryAllClubsService(highSchool: HighSchool): AllClubResponse
