+++ bitgouel-api/src/main/kotlin/team/msg/domain/student/presentation/data/response/StudentActivityResponse.kt
@@ -14,7 +14,7 @@ data class StudentActivityResponse(
     val approveStatus: ApproveStatus
 ) {
     companion object {
-        fun pageOf(studentActivities: Page<StudentActivity>,user: User): Page<StudentActivityResponse> =
+        fun pageOf(studentActivities: Page<StudentActivity>, user: User): Page<StudentActivityResponse> =
             studentActivities.map {
                 StudentActivityResponse(
                     activityId = it.id,
